refactor(core): start maintenance/ split — move shared types to maintenance/types.ts
First step of the maintenance.ts god-file split (tracked under
codemem-ug38). Extracts the 18 exported interfaces + MemoryRole string
union from the top of packages/core/src/maintenance.ts into
packages/core/src/maintenance/types.ts — the MemoryRole report types
and raw-event relink types cluster together logically and are referenced
from many callers.

maintenance.ts re-exports all 18 names so external consumers
(packages/cli/src/commands/memory.ts, packages/viewer-server/src/routes,
plus tests) keep working unchanged. Internal uses import directly from
the new types module.

Mechanical copy — no interface edits, no behavior change, no function
bodies touched. Gate stays green (tsc / lint / 1417 tests).
maintenance.ts drops from 2806 to 2602 lines.

Left for follow-up sub-commits on the same refactor bead:
 - status.ts (getRawEventStatus, rawEventsGate, retryRawEventFailures)
 - memory-role.ts (getMemoryRoleReport, compareMemoryRoleReports, probes)
 - relink.ts (getRawEventRelinkReport, getRawEventRelinkPlan, apply)
 - init-vacuum.ts (initDatabase, vacuumDatabase)
 - reliability.ts (getReliabilityMetrics + ReliabilityMetrics type)
 - tags.ts (backfillTagsText + types)
 - low-signal.ts (deactivateLowSignalObservations/Memories)
 - dedup.ts (dedupNearDuplicateMemories)
 - narrative.ts (backfillNarrativeFromBody, extractNarrativeFromBody)
 - dedup-keys.ts (backfillMemoryDedupKeys, plan, apply)
 - ai-structured.ts (aiBackfillStructuredContent)

Each follow-up will verbatim-copy the originals rather than rewriting.
